Create README.md
Project description
The customer is the credit department of the bank. It is necessary to find out whether the marital status and the number of children of the client impact the fact of repaying the loan on time. Initial data from the bank - statistics on the economic condition of customers.
The results of the study will be taken into consideration when building a credit-scoring model - a special system that assesses the ability of a potential borrower to repay a loan to a bank.
Data Description
children — number of children in the family
days_employed - total work experience in days
dob_years - client's age in years
education — the level of education of the client
education_id — education level identifier
family_status - marital status
family_status_id - marital status identifier
gender - gender of the client
income_type - type of employment
debt - whether he had debt to repay loans
total_income - monthly income
purpose - the purpose of obtaining a loan